# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Property Search API",
        "description": "API for searching properties",
        "contact": {
            "name": "John Doe",
            "email": "properties@properties.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/properties/search": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Search properties",
                "operationId": "84d21be2b9e1a169935c989ed89663c0",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name to search",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "bedrooms",
                        "in": "query",
                        "description": "Number of bedrooms",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bathrooms",
                        "in": "query",
                        "description": "Number of bathrooms",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "storeys",
                        "in": "query",
                        "description": "Number of storeys",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "garages",
                        "in": "query",
                        "description": "Number of garages",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "prices",
                        "in": "query",
                        "description": "price range",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        }
    }
}